Style/AndOr should have EnforcedStyle: conditional
Running `standardrb --fix` should only be making safe changes and `Style/AndOr` with `EnforcedStyle: always` is not safe. A common example is in controllers when you want to return early after a render or redirect

```
  redirect_to some_path and return
```

with `EnforcedStyle: always` it will autocorrect it to
```
  redirect_to some_path && return
```
which does not do the same thing. 

https://www.rubydoc.info/gems/rubocop/RuboCop/Cop/Style/AndOr